requirements.txt
amqp==5.1.1
asgiref==3.6.0
async-timeout==4.0.2
attrs==22.2.0
certifi==2022.12.7
cffi==1.15.1
coreschema==0.0.4
cryptography==38.0.4
daphne==3.0.2
defusedxml==0.7.1
dj-database-url==0.5.0
Django==4.1
django-auditlog==2.3.0
django-celery-results==2.5.0
django-cloudinary-storage==0.3.0
django-configurations==2.2
django-cors-headers==3.13.0
django-csp==3.7
django-jazzmin==2.6.0
django-debug-toolbar==3.2.1
django-extensions==3.1.3
django-extra-fields==3.0.2
django-filter==23.3
django-hashids==0.7.0
django-phonenumber-field==7.1.0
django-phonenumbers==1.0.1
django-storages==1.13.2
django-templated-mail==1.1.1
djangorestframework==3.14.0
djangorestframework-simplejwt==4.8.0
djoser==2.1.0
drf-yasg==1.21.4
grpcio-status==1.51.3
gunicorn==20.1.0
Jinja2==3.1.2
jmespath==1.0.1
kombu==5.2.4
MarkupSafe==2.1.1
mccabe==0.7.0
msgpack==1.0.5
oauthlib==3.2.2
packaging==22.0
phonenumbers==8.13.20
Pillow==9.4.0
prompt-toolkit==3.0.38
proto-plus==1.22.2
protobuf==4.22.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.10.0
pycparser==2.21
pycryptodome==3.18.0
pyflakes==3.0.1
PyJWT==2.6.0
pyOpenSSL==23.0.0
pyparsing==3.0.9
PyPDF2==3.0.1
python-dateutil==2.8.2
python-dotenv==0.21.0
python3-openid==3.2.0
requests==2.28.1
requests-oauthlib==1.3.1
whitenoise==6.7.0
